error handling for file types

+++ routes/index.js
@@ -4,6 +4,7 @@ var jimp = require('jimp');
 var multer = require('multer');
 var upload = multer({dest: "./public/uploads/"});
 var fs = require('fs');
+var url = require('url');
 
 /* GET home page. */
 router.get('/', function(req, res, next) {
@@ -11,6 +12,19 @@ router.get('/', function(req, res, next) {
 });
 
 router.post('/alright', upload.single("file"), function(req, res, next) {
+  var mime = /image\/.*/,
+      sendObj = {},
+      origList = req.file.originalname.split('.');
+
+  if(req.file.mimetype.match(mime) === null){
+      sendObj.fileName = 'pitythefool.html';
+      sendObj.originalName = 'pitythefool';
+      return res.send(sendObj);
+  }
+
+  origList.pop();
+  sendObj.originalName = origList.join('');
+
   jimp.read(req.file.path, function(err, image){
       if(err) res.send(err);
       var currentRow = 0,
@@ -35,17 +49,22 @@ router.post('/alright', upload.single("file"), function(req, res, next) {
       htmlImg = wrapStart+imgMarkup+wrapEnd;
 
       fileName = 'img_'+ Date.now() +'.html'
+      sendObj.fileName = fileName;
       fs.unlink(req.file.path);
       fs.writeFile('./public/downloads/' + fileName, htmlImg, function(err){
-          res.send(fileName);
+          res.send(sendObj);
       });
   });
 });
 
 router.get('/yeah/:path', function(req, res){
-    res.download('./public/downloads/' + req.params.path, 'imageYEAH.html', function(err){
+    var url_parts = url.parse(req.url, true);
+    var query = url_parts.query;
+    res.download('./public/downloads/' + req.params.path, query.originalName + '-rasterfyed.html', function(err){
         console.log(err);
-        fs.unlink('./public/downloads/' + req.params.path);
+        if(req.params.path !== 'pitythefool.html'){
+            fs.unlink('./public/downloads/' + req.params.path);
+        }
     });
 });
